Create a util module and function to call yaml.safe_load
There are multiple places where yaml.safe_load is called and the
previous patch missed those places. Create a common function under
orquesta.utils where the custom constructor is added to SafeLoader and
refactor codes that invokes yaml.safe_load to use the util function
instead.
