+++ orquesta/tests/unit/specs/native/test_workflow_spec_validate.py
@@ -1,3 +1,4 @@
+# Copyright 2021 The StackStorm Authors.
 # Copyright 2019 Extreme Networks, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,8 +15,9 @@
 
 import six
 
-from orquesta import exceptions as exc
+from orquesta.specs import loader as spec_loader
 from orquesta.tests.unit.specs.native import base as test_base
+from orquesta.utils import specs as spec_util
 
 
 class WorkflowSpecValidationTest(test_base.OrchestraWorkflowSpecTest):
@@ -489,11 +491,23 @@ class WorkflowSpecValidationTest(test_base.OrchestraWorkflowSpecTest):
                     do: task1
         """
 
+        expected_error = 'Failed to load workflow definition because found duplicate key "task1"'
+
         assertRaisesRegex = self.assertRaisesRegex if six.PY3 else self.assertRaisesRegexp
 
         assertRaisesRegex(
-            exc.WorkflowInspectionError,
-            "Workflow definition failed inspection.",
-            self.instantiate,
+            ValueError,
+            expected_error,
+            spec_util.instantiate,
+            self.spec_module_name,
+            wf_def,
+        )
+
+        spec_module = spec_loader.get_spec_module(self.spec_module_name)
+
+        assertRaisesRegex(
+            ValueError,
+            expected_error,
+            spec_module.instantiate,
             wf_def,
         )
