+++ orquesta/utils/specs.py
@@ -1,3 +1,4 @@
+# Copyright 2021 The StackStorm Authors.
 # Copyright 2019 Extreme Networks, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,57 +15,11 @@
 
 import logging
 import six
-import yaml
 
-try:
-    from yaml import CLoader as Loader
-except ImportError:
-    from yaml import Loader
-
-from yaml import constructor
-from yaml import nodes
-
-from orquesta import exceptions as exc
 from orquesta.specs import loader as spec_loader
-
+from orquesta.utils import yml as yaml_util
 
 LOG = logging.getLogger(__name__)
-
-
-# Custom YAML loader that throw an exception on duplicate key.
-# Credit: https://gist.github.com/pypt/94d747fe5180851196eb
-class UniqueKeyLoader(Loader):
-    def construct_mapping(self, node, deep=False):
-        if not isinstance(node, nodes.MappingNode):
-            raise constructor.ConstructorError(
-                None, None, "expected a mapping node, but found %s" % node.id, node.start_mark
-            )
-        mapping = {}
-        for key_node, value_node in node.value:
-            key = self.construct_object(key_node, deep=deep)
-            try:
-                hash(key)
-            except TypeError as exc:
-                raise constructor.ConstructorError(
-                    "while constructing a mapping",
-                    node.start_mark,
-                    "found unacceptable key (%s)" % exc,
-                    key_node.start_mark,
-                )
-            # check for duplicate keys
-            if key in mapping:
-                raise constructor.ConstructorError('found duplicate key "%s"' % key_node.value)
-            value = self.construct_object(value_node, deep=deep)
-            mapping[key] = value
-        return mapping
-
-
-# Add UniqueKeyLoader to the yaml SafeLoader so it is invoked by safe_load.
-yaml.add_constructor(
-    yaml.resolver.BaseResolver.DEFAULT_MAPPING_TAG,
-    UniqueKeyLoader.construct_mapping,
-    Loader=yaml.SafeLoader,
-)
 
 
 def instantiate(spec_type, definition):
@@ -72,12 +27,7 @@ def instantiate(spec_type, definition):
         raise ValueError("Workflow definition is empty.")
 
     if isinstance(definition, six.string_types):
-        try:
-            definition = yaml.safe_load(definition)
-        except constructor.ConstructorError as yex:
-            raise exc.WorkflowInspectionError([yex])
-        except Exception as ex:
-            raise ex
+        definition = yaml_util.safe_load(definition)
 
     if not isinstance(definition, dict):
         raise ValueError("Unable to convert workflow definition into dict.")
